Functioning Game V1.0

+++ tp1/gameElements/SlayerList.java
@@ -19,16 +19,21 @@ public class SlayerList {
         sList = new Slayer[tempSize];
     }
 
-    public void addSlayer(Game game, int posX, int posY){
+    public boolean addSlayer(Game game, int posX, int posY){
 //        sList.add(s);
-        if(!slayerHere(posX,posY)) {
-            sList[index++] = new Slayer(game, posX, posY);
+        if(game.getDimX()-1 != posY && posY != -1) {
+            if (!slayerHere(posX, posY)) {
+                sList[index++] = new Slayer(game, posX, posY);
+                return true;
+            }
         }
+
+        return false;
     }
 
     private boolean slayerHere(int x, int y) {
         for (Slayer s: sList) {
-            if(s.getPosX() == x && s.getPosY()==y){
+            if(s != null && s.getPosX() == x && s.getPosY()==y){
                 return true;
             }
         }
@@ -38,19 +43,22 @@ public class SlayerList {
 
     public void receiveDamage(int posX, int posY, int damage) {
 
-        for (int i = 0; i <= index; i++) {
-            if(nextTo(sList[i].getPosX(),sList[i].getPosY(),posX,posY)){
+        for (int i = 0; i < index; i++) {
+            if(sList[i]!= null && nextTo(sList[i].getPosX(),sList[i].getPosY(),posX,posY)){
                 sList[i].receiveDamage(damage);
             }
         }
     }
 
     public void removeDeadSlayers(){
-        for (int i = 0; i <= index; i++) {
-            if(!sList[i].isAlive()){
+        int tempI = 0;
+        for (int i = 0; i < index; i++) {
+            if(sList[i] != null && !sList[i].isAlive()){
                 sList = deleteSlayer(sList,i);
+                tempI++;
             }
         }
+        index = index - tempI;
     }
 
     private Slayer[] deleteSlayer(Slayer[] s , int i){
@@ -67,8 +75,10 @@ public class SlayerList {
 
 
     public void attack() {
-        for (int i = 0; i <= index; i++) {
-            sList[i].attack();
+        for (int i = 0; i < index; i++) {
+            if(sList[i] != null) {
+                sList[i].attack();
+            }
         }
     }
 
@@ -78,8 +88,8 @@ public class SlayerList {
     }
 
     public String slayerToString(int x, int y){
-        for (int i = 0; i <= index; i++) {
-            if(samePos(sList[i].getPosX(), sList[i].getPosY(), x, y)){
+        for (int i = 0; i < index; i++) {
+            if(sList[i] != null && samePos(sList[i].getPosX(), sList[i].getPosY(), x, y)){
                 return sList[i].toString();
             }
         }
@@ -87,13 +97,15 @@ public class SlayerList {
     }
 
 
-    public boolean isOnLeft(int posX, int posY) {
-        for (int i = 0; i <= index; i++) {
-            if(nextTo(sList[i].getPosX(),sList[i].getPosY(),posX,posY)){
-                return true;
+    public boolean isNotOnLeft(int posX, int posY) {
+        for (int i = 0; i < index; i++) {
+            if(sList[i] != null) {
+                if (nextTo(sList[i].getPosX(), sList[i].getPosY(), posX, posY)) {
+                    return false;
+                }
             }
         }
-        return false;
+        return true;
     }
